refactor: move errors related to Node compat to ext/node
